scaling/requirements.txt
motor==3.3.2
kafka-python==2.0.2
pymongo==4.6.0
pandas==2.1.4
pyarrow==14.0.2
minio==7.2.0
prometheus-client==0.19.0
asyncio-mqtt==0.13.0
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0